requirements.txt
Flask==3.0.0
flask-cors==4.0.0
numpy==1.26.4
Pillow==10.0.0
opencv-python==4.10.0.84
matplotlib==3.8.2